decoder: make sync::Support::doptions a Unit::DOption
Some time ago, we introduced the `unit::Unit` and `unit::IOptions`
traits for the handling of tracing unit implementation specific
`ioptions` in support packets. However, the `doptions` in those packets
are also implementation specific.

Previously, we would just read four bits (the size of the `doptions`
field for the reference flow) and store them in an `u64`. Given that the
field is at the very end of a packet, we could also have read 64 bits
without that making much of a difference.

However, in the long term we will probably want to handle those options
correctly. And even if we never implement data tracing ourselves, users
of this library could do so given meaningful `doptions`. We thus decided
to give `doptions` much of the same treatment as `ioptions`.