Clean up package metadata
We prefer a certain ordering for some fields and usually omit fields
that are infered (such as `readme`). We don't (yet) have a public
repository and while the library is _usable_ on an embedded system and
usualy traces programs _running_ on one, the `embedded` category is
meant for libraries that are primarily useful as a component for
software running on embedded devices.