instruction: add group of Kind fns for more convenient construction
We sometimes have to construct `Kind` values in code, particularly in
tests. For those variants that carry a `format::Type*`, construction
often was a bit awkward, and more than that, lengthy. We want to
condense tests more and thus decided to introduce options for shorter
construction of `Kind` values.

This does, of course, end up adding quite a few more lines to the
`instruction.rs` file, but we believe the benefits elswehre are worth
it.