instruction: add macro for generating instruction decode tests from data
The existing instruction decode tests are quite repetitive. And it will
become worse in the future if we add support for instructions not based
on RV32I. We therefore decided to introduce a macro which assembles a
testing fn based on a "raw" instruction, its decoded `Kind` and a few
properties we can query.