instruction: add enum representing base instruction set in dedicated mod
We currently only support decoding RV32I instructions (including
extensions defined for RV32I), because encodings may differ between base
instruction sets. However, we _do_ want to support other base sets. We
thus need to allow users to specify the base set when decoding an
instruction.

This change introduces a new enum representing base instructions with
the single variant (for RV32I) we currently support. We chose to place
it in a dedicated submodule in the hope of reducing the size of
`instruction` as future changes move some of the instruction decode
functionality over to this new submodule.