doc: add script running the simple example with reference flow data
`doc/development.md` describes in detail how to use the simple example
for testing against the reference flow. The description is close to a
script already, and now that we expect (nearly) all of those tests to
pass, we may as well provide a script we can lazily throw at the testing
data.