tracer: advance PC by correct size in state::State::exception_address
`state::State::exception_address` may need to adance the PC to the
location of an instruction that was not retired (but excepted).
Previously, we would advance it by the size of the instruction we were
at before calling `state::State::next_pc`, but we obviously need to
advance it by the size of the insn at the _current_ location.