tracer: make state::State::next_pc return PC and insn discreetly
We plan to redefine `Item`, distinguishing between `Item`s signalling
retired `Instruction`s and `Item`s signalling traps. The latter will not
carry an `Instruction`, which we may need when determining the EPC. So
it does not make that much sense to make `state::State::next_pc`
assemble an `Item` in the future. Since it already returns a tuple, we
may as well make that a 3-tuple instead of a pair.