tracer: make item::Item::instruction return an Option
Previously, we would report traps by tapering them on existing `Item`s.
This lead to some unclarities around the relation ship of the trap info
to the PC and `Instruction` reported by the `Item`. We thus came to the
conclusion that traps do deserve their very own, distinct `Item`s.

We recently introduced an `item::Kind`, which we will use for redefining
`Item`s. This change prepares for that redefinition by adapting the
`item::Item::instruction` query fn to `Item`s that may not carry an
`Instruction`.