tracer: make item::Item::trap only return the trap::Info
Previously, we would report traps by tapering them on existing `Item`s.
This lead to some unclarities around the relation ship of the trap info
to the PC and `Instruction` reported by the `Item`. We thus came to the
conclusion that traps do deserve their very own, distinct `Item`s.

We recently introduced an `item::Kind`, which we will use for redefining
`Item`s. This change prepares for that by adapting `item::Item::trap`:
the EPC will be signalled via `item::Item::pc`, and only for exceptions
as the trace does not provide an actual EPC for interrupts. This change
also adapts documentation that was never accurate in this respect, and
stopped being correct in others.