tracer: redefine item::Item
Previously, we would report traps by tapering them on existing `Item`s.
This lead to some unclarities around the relationship of the trap info
to the PC and `Instruction` reported by the `Item`. We thus came to the
conclusion that traps do deserve their very own, distinct `Item`s.

We recently introduced an `item::Kind`, which we now use for redefining
`Item`s. This affrects the public interface considerably, but also makes
things more clear around traps and probably also reduces the overall
size of an `Item`.