examples: add module for processing spike traces
The simple example has functionality for testing the library against
trace data generated through the reference flow. To be more precise it
compares the tracer output to a reference provided in the form of a CSV.

While the existing reference logic does its job most of the time, the
data produced around traps didn't come out right. We could try to fix
the existing logic, but it's already somewhat messy and distracts from
what the example intends to demonstrate.

We therefore decided to move the logic to its own module that we then
import in the example. This change thus introduces a new module with
clean logic for generating proper `Item`s from the reference that can be
compared to the `tracer::Tracer`'s output as is.