examples: use new spike::CSVTrace for reference
The simple example has functionality for testing the library against
trace data generated through the reference flow. To be more precise it
compares the tracer output to a reference provided in the form of a CSV.

While the existing reference logic does its job most of the time, the
data produced around traps didn't come out right, and comparison to
`Item`s was awkward. We therefore recently introduced a dedicated module
with code for generating a proper sequence of `Item`s from the reference
CSV.

This change migrates the simple example to that logic.