examples: make last insn in simple example's bootrom a jalr
We use the simple example for rough testing against the reference flow.
The traces we get from that also include the initial bootrom, thus we
also include it in the example so we don't stumble over a program
section that's not in the ELF file. However, while we did duplicate the
bootrom's behaviour, we noted down a `c.jr` instead of a `jalr`. Which
will lead to failures should we start also comparing the actual
`Instruction` reported.