tracer: add type representing the Kind of an Item
Previously, we would report traps by tapering them on existing `Item`s.
This lead to some unclarities around the relation ship of the trap info
to the PC and `Instruction` reported by the `Item`. We thus came to the
conclusion that traps do deserve their very own, distinct `Item`s.

We intend to have an `Item` always carry a PC, which will either be the
PC of a retired `Instruction` or a PC associated to a trap (e.g. the PC
of the instruction that trapped). However, whether an `Item` carries an
`Instruction` or a `trap::Info` will depend on the kind of `Item`. This
change introduces a representation of that `Kind` of `Item`.

Future changes will redefine `Item` in terms of that `Kind`.