tracer: add state::State fns for querying the current PC and insn
We plan to redefine `Item`, distinguishing between `Item`s signalling
retired `Instruction`s and `Item`s signalling traps. This will make the
current use of `state::State::current_item` for determining an EPC
awkward.

We therefore decided to introduce a fn for querying the PC. We also
include a fn for querying the current insn, which allows us to move the
responsibility of creating an `Item` away from `state::State` to
`Tracer` in the future.