lib: enable unstable feature doc_auto_cfg for docs.rs
We want all features visible in public docs, but we also want items only
availible for some feature to be marked. The unstable `doc_auto_cfg`
feature does what we want, but it is... unstable and therefore only
availible on nightly toolchanins.

We know that https://docs.rs runs a nightly version of rustdoc, so we
can at least activate this feature there.