examples: change output for individual trace Items in simple example
The example's output was originally meant to resemble the output of the
reference flow's decoder model. The idea was to find divergence between
our tracer and the reference model by simply `diff`ing the output.

However, we since added functionality to compare trace `Item`s directly
to the spike trace from which the instruction trace was produced. We no
longer resort to `diff`ing textual outputs. Thus, we do have greater
freedom in what the example prints to stdout.

This change exercises this freedom and makes the example print one line
per `Item` in a more or less human friendly but compact format.