instruction: add tests::DecodeForTest fn returning an Option
... rather than panicing.

Currently, we test whether known encodings result in the instructions we
expect. However, we do not (yet) ensure we don't errorneously decode
"random" data into an instruction it isn't. This requires a decoding
method that does not panic but lets us assert that the data decodes to
`None` wrt. `Kind`.