instruction: fix segment filtering in elf::Elf::get_insn
In order to find the `Instruction` at the requested address,
`elf::Elf::get_insn` iterates over all segments and filters them. We
use `slice::checked_split_at` to remove any part of a segment before the
instruction and filter at the same time. However, we previously would
not ensure that the resulting slice was non-empty. Thus, with two
consecutive segments, for requests for the first address in the second
segment, we may have ended up selecting the first one and fail to decode
any `Instruction` (from the empty slice).

This change resolves this issue, slightly restructuring the code around
the affected filter step.