instruction: do not store last section in an Option in elf::Elf
`elf::Elf` has an optimization for repeated queries of `Instruction`s
from one segment: we cache the last segment from which data was
requested. Previously, we would store the respective data in an
`Option` which was set to `None` on creation, since we did not yet
identify a relevant segment to cache. However, this did penaltize any
segment search but the first (successful) one.

We therefore decided to _not_ use an `Option`. Instead, we initialize
the cache with a non-existent empty section at `u64::MAX`. Naturally,
any queries would ignore that section and eventually, it would be
updated to an existing one.