instruction: add Display trait for instruction::Kind
To show which instruction is executed, the Display trait is now
implemented for for instruction::Kind. This allows for showing the name
of the executed function. Instructions with specified format type will now
print the name of their instruction as well as their register or
immediate values, while instructions of unknown format only display
their name.