instruction: add trait allowing binary errors to signal misses
A trace may cover code supplied via different `Binary`s. The reference
flow's traces, for example, cover the bootrom, the application and in
some cases a firmware. We therefore need to combine them `Binary`s.
Usually, a given `Binary` will cover one or more memory areas.

This change introduces a trait for differentiating between errors
returned by `Binary::get_insn` because the address is not in one of its
memory areas and actual failures. This will allow querying multiple
`Binary`s and returning the first response that's not a miss —even if it
is an error.

The trait also features a fn for creating such an error, which allows
generation of a proper error for cases where all options (in the form of
`Binary`s) are exhaused.