instruction: add slice-like compond binary::Binary
We may need to handle multiple `binary::Binary`s of the same type in the
future, e.g. an ELF each for firmware and application. And in some
cases, the number is not fixed.

In such a situation, it seems natural to iterate over the ELFs and yield
the first result which is not a "miss", using the recently introduced
`MaybeMiss`. This change introduces an utility for doing so,
encapsulating a slice and doing some basic caching wrt. the `Binary`
selection.