types: make trap::Info::ecause an u16
Previously, we would use an `u64` for `ecause` values. However, in
practice ecauses are much smaller: the standart currently only defines
ecauses up to `64`, and while platforms may introduce higher ecauses for
interrupts, it is rather unlikely that they would require 64 bits.
Looking at "typical" values for the `ecause_width_p` parameter, 8 bits
are probably enough.

The size of the `ecause` field has a significant impact on the size of
`trap::Info` and therefore `tracer::Item`, which gets moved a lot. We
therefore decided to use a smaller unsigned integer type. We decided on
an `u16` rather than an `u8` just to be an the safe side. We can afford
an extra byte.