instruction: change to Display for instruction::format in commit 634dd147b9d861fa821fbd6c6ea80b0eb74ee328
Beforehand the register values for the format types were displayed,
however doing this it was unclear if the value displayed was an immediate or
a register. For better classification as registers in the
display trait an 'x' was added right before the register value,
identifiying it as e.g. x4 and not 4. The immediates are now displayed
as hexa decimal values for clarity, in format of '0x'immediate_value.