tracer: add struct representing an execution context
We plan to communicate the execution context to consumers of the
`Tracer`'s `Iteratior` interface via a special `item::Item` type. We
could just introduce a variant with privilege level and context, but a
type that implements `Eq` and some other traits for the execution
context has its own merit. Especially considering we'll not _only_
communicate actual context changes but any context we receive.