tracer: redefine state::StopCondition::Sync
We already communicate some updates to the execution context via
`Item`s. `Tracer` emits those for all cases that could be handled
entirely via the `Tracer`'s `IterationState`, without involving
`state::State`.

The remaining case is an context update for the last (regular) `Item`
emitted when the `state::State` hits the synchronization stop condition.
Previously, such a `state::StopCondition::Sync` would ony contain a
`Privilege`, and only if it was relevant for the condition, i.e. when it
would control whether or not the execution needed to be followed
further (in version 1 of the trace protocol).

In order to generate a context update `Item`, we need the full
`item::Context`, and we need it regardless of whether or not we are
confronted with a version 1 or version 2 trace. At the same time, we
still need the information what action to take when hitting the stop
condition, i.e. whether to also compare the privilege level or update
the context.