tracer: move privilege update out of Tracer::sync_init
We already communicate some updates to the execution context via tracing
`Item`s. `Tracer` emits those for all cases that could be handled
entirely via the `Tracer`'s `IterationState`, without involving
`state::State`. We recently redefined `state::StopCondition::Sync` in
order to handle the remaining case. It now always carries an
`item::Context` and updates the privilege level in `state::State` when
not operating in trace version 1 mode. Thus, we don't want to update it
for this particular type of sync packet.

Since we want to keep using `Tracer::sync_init` we thus need to move the
privilege level update ouf of that fn.