tracer: add item::Kind variant for context changes
We plan to communicate the execution context to consumers of the
`Tracer`'s `Iteratior` interface via a special `item::Item` type. We
could just introduce a variant with privilege level and context. Instead
we introduced a dedicated type representing the execution context. This
change allows transporting values of that type in an `item::Kind` and
thus an `item::Item`.