tracer: add IterationState variant for context changes
We plan to communicate updates to the execution context via `Item`s. For
that purpose we recently introduced a new type representing a context
and an `item::Kind` variant transporting them.

Context updates are single events introduced by synchronization packets.
For some packet types the update is immediate and, like traps, an update
may or may not be sollowed by a single item —which is currently emitted
without the context update.

This change introduces an `IterationState` variant which let us inject
context items. Future changes will produce this variant.