tracer: inject context item on immediate context change
We want to communicate updates to the execution context via `Item`s.
This change does so for any contet update that is "immediate", meaning
a sync packet results in a context `Item` to be emitted before any
regular `Item`. These emissions are controlled entirely by the
`Tracer`'s `IterationState`, i.e. `state::State` is not involved.