examples: make simple example ignore missmatches due to sync packets
We plan to communicate updates to the execution context via
`tracer::Item`s. For that purpose we recently introduced a new type
representing a context and a `tracer::item::Kind` variant transporting
them.

A real trace will contain sync packets that are not reflected in current
`tracer::Item`s but will be visible as context updates in the future.
And trying to generate the exact same ones in the `spike` helper module
is not something we want to do. Instead, we introduce some logic for
ignoring sync packets.