examples: make simple example only abort on mismatches for Item::Regulsr
Originally, the "simple" example would abort in reference mode if an
`Item` differed in any way from a reference `Item`. We relaxed this in

    65fd5ff99f931be3488087ca9321625af4bda278

and only aborted if the PC differed. This change now relaxes this
condition further, making the "simple" example only abort on PC
mismatches for regular items, as those are the only ones which result in
diverging execution paths.

We still do report any mismatch and consider them worth investigation.