tracer: integrate handling of end in handling of StopCondition
Previously, we would handle the `state::StopCondition` and the `end`
flag returned by `state::State::next_pc` roughly like the reference
implementation does, with the notable difference that the reference
does not have the concept of a `state::StopCondition` but did all that
based on sifting through packet data directly.

Since we set the stop condition to `state::StopCondition::Fused` if the
`end` flag was set, we'd mask the previous stop condition and thus not
hit any case (but the default one) in the `match` statement handling
specific stop conditions. This is both inefficient (since the match is
still executed) and masks logic we'll want to introduce there in the
future.

We therefore decided to move the `end` handling into the match, as the
very last case before the default one. This preserves current behaviour
because any existing case handling a specific `state::StopCondition` can
only be active if we're inside the `branch_limit`.