decoder: add macros for generating decoder tests
We have two types of tests in `decoder`: tests for specific `Decoder` fn
and tests of `Decode` impls. The latter are both uniform and contain
some boilerplate. And if we ever introduce encoding capabilities, we'll
want an encoding test for each of those test cases.

We threfore decided to express those tests more concisely via a macro.
If we introduce an encoder, we can simply tweak the macro itself to
produce not just a decoder but also an encoder test. Because of this
likely future use we choose the somewhat unusual name `bitstream_test`
as we check whether a value can be de—/encoded correctly from a given
bitstream.