decoder: remove len member from smi::Packet
Originally, we needed to convey the sice of a packet to callers of
`Decoder::decode_smi_packet` because they needed to discard the part of
the input that was decoded. We since changes the `Decoder` interface
significantly, and it is no longer necessary to explicitly retrieve the
length of a single `smi::Packet`. On the other hand, the member gets in
the way of future plans to support _encoding_ packets, as the length is
a result of the encoding and not an input.