decoder: replace smi::Packet::header with individual values
Previously, `smi::Packet` would consist of an `smi::Header`, with its
own `Decode` impl and a payload, also with a `Decode` impl. The header
would exclusively be used as part of, and its semantics tightly coupled
to, an `smi::Packet`. Users would need to explicitly access the header
for data they needed for operations, such as the hart index, while they
would plug the payload into the appropriate `Tracer` as is.

We feel this is unnecessary and decided to embed the necessary
information directly into `smi::Packet`. We dropped `trace_type` since
it has a single legal value. If it gained another one, we'd map that to
different payloads via an appropriate enum. We also dropped the payload
length since that's of no consequence for operations and stands in the
way of possible future encoding functionality. Finally, we made the hart
index an `u64`, since it's neither a size or an slice index, and renamed
it to `hart` because we feel the "index" is redundant.