decoder: add Decoder fn for retrieving the current byte position
We plan to provide more `Decode` impls, e.g. for `smi::Packet`. Some of
those impls will need to query the current byte position, and we'd like
to do that via a fn rather than accessing `Decoder::bit_pos` directly.
And since the fn is probably at least of some interest to some users of
this library, we may as well make it public.