decoder: add fn for restricted decoding
We plan to impl `Decode` for `smi::Packet` and potentially other
"packet" types in the future. For that, we need the ability to decode a
payload that may be compressed, i.e. we need to restrict what bytes may
be extracted while decoding an item.

The existing `Decoder::decode_smi_packet` already contains logic to do
that, but that requires access to `Dercoder` internals. We therefore
decided to introduce a new fn.