decoder: make Decoder::bytes_left account for non-zero bit positions
`Decoder::bytes_left` was, and is, meant to be called by users of the
library before or after decoding packets in order to gauge haow much of
the trace data still needs to be processed. As decoding an SMI packet
would leave the `Decoder` in a reset state, with a `bit_pos` of zero, we
could just return the length of the internal data.

However, we also allow decoding a payload, which does not reset the
`Decoder`. And we may choose to support other formats in the future for
which we also don't do so. And besides, we do want the fn to return what
is advertised.