lib: add alloc feature pulling in the alloc crate
We made this library not depend on `std`, making it suitable for
embedded application. As a result, we did not use, or implemented our
traits for, collections or data structures that would require memory
allocation. Which would be particularly useful for dynamic dispatching
of `instruction::binary::Binary` or `decoder::unit::Unit`.

The `alloc` crate provides everything we need for that, but since some
users may not want that, we made it optional.