decoder: add PULP trace unit::Unit
Both for our users' sake and testing opportunities, we want to include
support for at least common tracing units. This change instroduces
support for the `rv_tracer` from the PULP ecosystem. Particularl since a
changed version with identical `ioptions` is in the process of being
integrated in the CVA6 repository.

Sadly, the original `rv_tracer` is not in a particularly good state, but
we at least got it to generate support packets which let us confirm that
we do decode these `ioptions` correctly.