binary: add Binary wrapper for boxing Errors
Users of this library may want the ability to plug and combine different
kinds of `Binary` into a `tracer::Tracer`. For example, they may want to
combine arbitraty ELF files and raw binaries. The only obstacle left to
a `dyn Binary` is `Error`, which typically differs.

This change introduces a `Binary` wrapper that wraps the `Error` type in
a `Box<dyn error::MaybeMissError>` which impls `error::Miss`. This
allows combining `Binary`s of different types via the existing utilities.