binary: add provided fn to Binary for creating a BoxedError
We recently introduced the `Binary` adapter `boxed::BoxedError` which
allows dynamically dispatching `Binary`s. This was the last missing
piece for combining `Binary`s with different `Error` types without the
need for users to introduce helper types of any sort. However, doing so
was not yet ergonomic, as a `BoxedError` needed to be created, boxed and
the `Box` then coerced into a specific, awkwardly complicated type.

This change allows doing all that with a simple call to a provided fn.
We made the explicit choice to allow for encapsulation of non-`'static`
`Binary`s because of `elf::Elf`. That `Binary` will be both quite common
in applications and be restricted by the lifetime attached to the
underlying `ElfBytes`.