binary: add a Binary wrapping an FnMut
We currently have a blanket impl of `Binary` for `FnMut`s. This impl can
be useful, but obstructs other very useful impls, e.g. for "reference"
types. We therefore decided to remove it. But we do want to preserve the
option to plug in a `Fn` as a `Binary`. This change introduces a small
wrapper for that purpose.