binary: make instruction::binary a top level module
The `binary` module already has a few utilities, and we plan to add
quite a few more. While the number of lines will probably remain
manageable, it feels there's too much mashed together. Since module is
already somewhat distinct in `instruction` to begin with, we decided to
make it a toplevel module. We also move `instruciton::elf` over as a
first submodule, as it should have been to begin with.