binary: add Binary for simple maps
Currently, `&[(u64, Instrction)]` impls `Binary`. This did serve its
original purpose, but it's also somewhat restrictive as it does not
enable using a `Vec` directly. We cannot and should not introduce a
blanket impl for all types implementing `AsRef` since that would
conflict with other impls.

Just like for `FnMut`, we thus decided to introduce a `Binary` adapter.
And sicne we control how it's created, we can enforce the slice being
sorted by address.