binary: impl Extend for Multi
In applications using `std`, `Multi` will often be used with a `Vec` or
similar container. That container may well be constructed using
`Extend::extend`. This impl allows creating a `Multi` early and then use
`Extend::extend` on it, rather than carrying around a `Vec` and
converting it to a `Multi` later.