binary: add trait combining error::MaybeMiss and core::error::Error
We plan to introduce impls and helper types for using `Binary`s with
dynamically dispatches, boxed errors. This will require writing out
types like `Box<dyn MaybeMiss>`. However, since we are dealing with
error types we _do_ want to make them `core::error::Error`. And we
cannot dynamically dispatch multiple non-auto traits.

Requiring `core::error::Error` for `MaybeMiss` is not an option as this
would introduce a requirement for code which does not use `alloc` at
all. Also, we do like the convenience of the `MaybeMiss` impl for
`Result`. So instead, we decided to introduce a new trait just for the
`Box<dyn>` use-case. We also include a blanket impl, removing the burden
to impl that trait for every single relevant tyoe.