binary: impl error::Miss for Box<dyn error::MaybeMissError>
Users of this library may want the ability to plug and combine different
kinds of `Binary` into a `tracer::Tracer`. For example, they may want to
combine arbitraty ELF files and raw binaries. The only obstacle left to
`dyn Binary` is `Binary::Error`, which typically differs. They could
just `Box` the `Error` (if the `alloc` feature is enables), but if we
want to combine `Binary`s we need it to also impl `error::Miss`.

We recently implementes `error::MaybeMiss` for `Box<error::MaybeMiss>`.
That impl also covers `Box<dyn error::MaybeMissError>`. This change now
also impls `error::Miss` for that type, i.e. it allows creation of a
"miss" even if none of the `Binary::Error`s do.