binary: move Multi to new combinators submodule
We recently made `binary` a top-level module and splitt of an `error`
submodule. However, the module continues to be convoluted as it contains
quite a few different items —mostly types that impl `Binary`.

This change introduces a new submodule for a category of `Binary` types,
namely `Binary` types that combine other `Binary`s, and moves `Multi` to
the new submodule. We do not move `Offset` since that doesn't _combine_
`Binary`s.