binary: move Func, from_fn and Empty to new basic submodule
We recently made `binary` a top-level module and splitt of an `error`
submodule. However, the module continues to be convoluted as it contains
quite a few different items —mostly types that impl `Binary`.

This change introduces a new submodule for a category of `Binary` types,
namely "simple" ones that are self-contained, or at least only adapt
some other non-`Binary` to serve as a `Binary`.