binary: remove Binary impl for &[(u64, Instruction)]
This impl did serve its original purpose, but it was also somewhat
restrictive as it does not enable using a `Vec` directly. We cannot and
should not introduce a blanket impl for all types implementing `AsRef`
since that would conflict with other impls.

Just like for `FnMut`, we thus introduced a `Binary` adapter and
migrated all existing users. This change now removes this impl.