examples: make inclusion of spike bootrom in "simple" example optional
We included the bootrom because we ran the example with traces generated
by the reference flow, which included these instructions. However,
traces found in the wild will usually not have the bootrom, and
especially not this particular one —albeit probably something very
sililar.

This change makes the "simple" example not include the spike bootrom in
the `binary::Binary` unless a specific command line flag is set. It
needs to be set in our reference tests.