examples: allow enabling simple example's debug output via CLI flag
Previously, we did not have a proper command line parser but still
wanted the ability to enable some additional debug output. We thus
decided to check an environment variable. Now that we have a proper
command line parser there is no reason not to have the command line
option. And we can even have the environment variable as an alternative.