binary: make offset not wrap
When we originally introduced `Offset`, we did not have `error::Miss` or
anything equivalent. Thus we were stuck with transformations that were
always successful, and therefore decided to make the address space wrap.

However, this is not what users usually expect, or want. And with `Miss`
we can easily handle addresses that are lower than the offset, albeit
requireing a new type bound.