decoder: add a unit::Unit for type erasing the specific unit::Unit
Previously, the choice of a `unit::Unit` for a `Decoder` needed to be
made at compile time, as the trait is not dyn-compatible. This change
introduces a helper allowing for run-time selection of a `unit::Unit`.

We decided to just fetch the `encoder_mode_width` once and store its
value because `unit::Unit`s are not supposed to change this value after
creation. Or at least while they are part of a `Decoder`.