decoder: add list of all unit::Units with constructors
Users of this library may want to allow choosing from all the
`unit::Unit`s we support in their applications. This is possible thanks
to `unit::Plug`, but would previously require them to update their code
whenever we would add a new `unit::Unit`.

This change (hopefully) makes their lifes easier by suppliying a list of
those with names and constructors. In their application, they can use
this list for menus or iterate over it looking for a user-supplied name.

We chose to make this a slice rather than an array because this way new
entries won't affect the lists's signature.