instruction: add basic support for 48bit and 64bit long instructions
The RISC-V Instruction Set Manual Volume I: Unprivileged Architecture
describes instruction length encodings past the 16bit and 32bit we dealt
with so far. While there are currently no ratified 48bit or 64bit
instruction encodings, there have been proposals and custom instructions
apparently exist.

Previously, we would stumble over them and generate an error. This
change allows us to properly recognise (and ignore) them. We obviously
assume that they don't contribute to the control flow in a way that
would be relevant for tracing.

We don't have any testing data on hand, but at least this addition is
guranteed to not interfere with decoding valid 16bit and 32bit
instructions.