examples: move complex reference checking logic from "simple" to spike
We do use the "simple" example for checking against a reference, but
that's not what we want to show there. Because of this we moved the
reference trace generation logic into its own submodule in

    a2a9c5bf71516e18b89be8ac36a5ba0021ea90dd

The logic for comparing these items, however remained in "simple" since
that used to be fairly small. However, it grew more complex again when
we introduced context items. We therefore decided to now also move that
piece of logic out of the example.