decoder: make Decoder::split_data set Decoder's data
Previously, `Decoder::split_data` would just return the two halves of
the data. Given that we use it for some form of restricted decoding,
we'll want to set the `Decoder`'s inner data to the first half right
afterwards anyway. Doing so and only returning the second half ends up
being a cleaner interface.