decoder: add fn for resetting a Decoder
The original `Decoder` design had a `reset` fn —that did not reset its
inner data because `Decoder` didn't carry any back then. Rather than
adapting it, we removed the fn in

    845a7be753329621dc46ffa2f763f56383aa49c1

However, we intend to introduce ephemeral packet types that allow
deferring or skipping decoding the payload. In both cases, we need to
advance the `Decoder` past the end of the packet, and we'll likely use
`Decoder::split_data` for them and thus have the data to reset the
`Decoder` to.