decoder: make Decode trait generic over Decoder and data lifetimes
The purpose of `Decode` is decoding packets and payloads, which are
currently all `'static` by design. Thus, `Decode` just took a reference
to a `Decoder` with arbitrary data lifetime.

While payloads will indeed always be `'static`, we intend to introduce
ephemeral packet types. These will allow defering or skipping payload
decoding. For that, they need to hold a reference to the `Decoder` and
potentially to data borrowed from it.

We could introduce regular, associated decode fn for those types, which
would declare the lifetimes needed. However, we'd loose the (future)
composability we would gain from `Decode` impls.