decoder: add type for normal encapsulation structure
We plan to introduce support for packets as described by the relatively
new Unformatted Trace & Diagnostic Data Packet Encapsulation for RISC-V
specification. This change introduces a type to represent a "Normal
Encapsulation Structure", which is one type of packet defined by the
encapsulation specification.

We implement this as an ephemeral packet with deferred payload decoding.
This allows users to skip decoding payloads that are of some different
type not covered by this libray, based on the source id and flow.

Decoding/instantiation of these data types will be performed through
another data type covering all possible encapsulation packet types. This
type will be introduced by a future change.