doc: make test-reference signal failed tests at the end
We want to run all the tests, and thus we did made the concious choice
not to abort if any one test failed. However, this also meant we had to
squint at the output to check whether any test failed. Ideally, we would
want to only need at the exit code.

This change allows for that by making the script exist with a non-zero
code and print a list of tests that failed.