decoder: impl PartialEq for unit::PULP and unit::PULPIOptions
The `PartialEq` impl for `unit::PULP` does offer little utility on its
own. But it allows us to simply derive `PartialEq` for types that are
generic over units and use that derived impl in tests.

we throw in an `PartialEq` impl for `unit::PULPIOptions` because _that_
is likely useful for some users.