decoder: rename payload::Payload to InstructionTrace
The current `payload::Payload` only covers instruction trace payloads.
We plan to introduce a new type that also covers data trace payloads in
some fashion. Naturally, we'd use the term "payload" for that one, which
makes things a bit confusing.

This change frees the name by renaming the current `payload::Payload` to
`InstructionTrace`.