decoder: move encap::Payload to payload
RISC-V encapsulation packets were the first for which we supported the
decoding of data trace payloads —in the sense that we'd identify these
payloads. We will likely also want this very same functionality for
other packet formats, which will result in very similar `Payload` types,
which would mean code duplication.

We therefore decided to make the `encap::Payload` _not_ specific to the
RISC-V packet encapsulation but more general.