decoder: add preliminary interface for ephemeral smi::Packets
We plan to make `smi::Packets` ephemeral, allowing users to decide
whether or not to decode its payload. With the new interface, the data
members of `smi::Packets` will not be public any more. Instead, we'll
use query fns. Introducing these now allows for a more gradual migration
of existing users.

This change introduces said query fns, including a fn which will, in the
future, be used for decoding the payload. Since we may want to allow
decoding data trace payloads in the future, we made the decoding fn
return a `decoder::payload::Payload`.